chore: move span utils to own module + implement non overlapping bilou
